Additional Shape Parameters and Plotting
Added the Direct and Indirect AMLS methods from Fasshauer and Zhang. Currently, the Indirect method does not perform perfectly.

The TEAD functions were grouped into a single adaptiveRefinement() function such that additionalAdaptive Refinement methods may be added in the future.

A plotting function was created that multiple simulation parameters, multiple methods, and multiple errors. However, in the current moment single simulation parameters have not been implemented.

A kwarg has been added to simulate which allows for directly indicating the shape_parameter, skipping the shape parameter selection process, allowing for a plot of error vs. shape parameter to be made. This was used for bug fixing purposes and may be removed later.

The adaptive refinement will now stop if there are no more points to evaluate, regardless of whether or not the tolerance has been met.

The Indirect AMLS method necessitates that the basis function is calculated with a previously determined 'dist', rather than one calculated with the points and centers. A new basis function was created for this purpose.

A verbose kwarg was added to allow or remove console logs. This will also be removed prior to merging with main branch.

The Kuo and Franke shape parameter methods were removed because they are only proven to be functional with MQ basis functions and the monotonicity of our solution is only guaranteed with Gaussian basis functions.

The timing of the simulation is also saved within the model struct.